fixes + add reducer

+++ src/modules/gift-search/reducers/filter.js
@@ -11,8 +11,28 @@ function filterReducer (state: FilterState = {}, action: Action): FilterState {
 	switch (action.type){
 		case "GIFT_LIST_SEARCH/SET_FILTERS":
 			return {...state, ...action.filters};
+
 		case "GIFT_LIST_SEARCH/RESET_FILTERS":
-			return deletePropertiesImmutable(state, action.filtersToReset);
+			const currentFilters = Object.keys(state);
+			let hasFilterToReset = false;
+			for (const filterToReset of action.filtersToReset) {
+				if(currentFilters.indexOf(filterToReset) > -1){
+					hasFilterToReset = true;
+					break;
+				}
+			}
+
+			if(hasFilterToReset)
+				return deletePropertiesImmutable(state, action.filtersToReset);
+			else
+				return state;
+
+		case "GIFT_LIST_SEARCH/RESET_ALL_FILTERS":
+			if(Object.keys(state).length === 0)
+				return state;
+			else
+				return {};
+
 		default:
 			return state;		
 	}
